+++ test/MaterialDatatable.test.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { spy, stub } from "sinon";
-import { mount, shallow } from "enzyme";
-import { assert, expect, should } from "chai";
+import {spy, stub} from "sinon";
+import {mount, shallow} from "enzyme";
+import {assert, expect, should} from "chai";
 import MaterialDatatable from "../src/MaterialDatatable";
 import MaterialDatatableFilterList from "../src/MaterialDatatableFilterList";
 import MaterialDatatablePagination from "../src/MaterialDatatablePagination";
@@ -9,124 +9,152 @@ import textLabels from "../src/textLabels";
 import Chip from "@material-ui/core/Chip";
 import Cities from "../examples/component/cities";
 
-describe("<MaterialDatatable />", function() {
+describe("<MaterialDatatable />", function () {
   let data;
   let displayData;
   let columns;
   let tableData;
   let renderCities = (value, tableMeta, updateValueFn) => (
-    <Cities value={value} index={tableMeta.rowIndex} change={event => updateValueFn(event)} />
+      <Cities value={value} index={tableMeta.rowIndex} change={event => updateValueFn(event)}/>
   );
-  let renderName = value => value.split(" ")[1] + ", " + value.split(" ")[0];
+  let renderName = (value) => {
+    return value.split(" ")[1] + ", " + value.split(" ")[0]
+  };
 
   before(() => {
     columns = [
-      { name: "Name", options: { customBodyRender: renderName } },
-      "Company",
-      { name: "City", options: { customBodyRender: renderCities } },
-      { name: "State" },
+      {name: "Name", field: "name", options: {customBodyRender: renderName}},
+      {name: "Company", field: "company"},
+      {name: "City", field: "city", options: {customBodyRender: renderCities}},
+      {name: "State", field: "state",},
     ];
     data = [
-      ["Joe James", "Test Corp", "Yonkers", "NY"],
-      ["John Walsh", "Test Corp", "Hartford", "CT"],
-      ["Bob Herm", "Test Corp", "Tampa", "FL"],
-      ["James Houston", "Test Corp", "Dallas", "TX"],
+      {name: "Joe James", company: "Test Corp", city: "Yonkers", state: "NY"},
+      {name: "John Walsh", company: "Test Corp", city: "Hartford", state: "CT"},
+      {name: "Bob Herm", company: "Test Corp", city: "Tampa", state: "FL"},
+      {name: "James Houston", company: "Test Corp", city: "Dallas", state: "TX"}
     ];
     // internal table data built from source data provided
     displayData = JSON.stringify([
       {
-        data: ["James, Joe", "Test Corp", renderCities("Yonkers", { rowIndex: 0 }), "NY"],
+        data: ["James, Joe", "Test Corp", renderCities("Yonkers", {rowIndex: 0}), "NY"],
         dataIndex: 0,
       },
       {
-        data: ["Walsh, John", "Test Corp", renderCities("Hartford", { rowIndex: 1 }), "CT"],
+        data: ["Walsh, John", "Test Corp", renderCities("Hartford", {rowIndex: 1}), "CT"],
         dataIndex: 1,
       },
       {
-        data: ["Herm, Bob", "Test Corp", renderCities("Tampa", { rowIndex: 2 }), "FL"],
+        data: ["Herm, Bob", "Test Corp", renderCities("Tampa", {rowIndex: 2}), "FL"],
         dataIndex: 2,
       },
       {
-        data: ["Houston, James", "Test Corp", renderCities("Dallas", { rowIndex: 3 }), "TX"],
+        data: ["Houston, James", "Test Corp", renderCities("Dallas", {rowIndex: 3}), "TX"],
         dataIndex: 3,
       },
     ]);
     tableData = [
-      { index: 0, data: ["James, Joe", "Test Corp", renderCities("Yonkers", { rowIndex: 0 }), "NY"] },
-      { index: 1, data: ["Walsh, John", "Test Corp", renderCities("Hartford", { rowIndex: 1 }), "CT"] },
-      { index: 2, data: ["Herm, Bob", "Test Corp", renderCities("Tampa", { rowIndex: 2 }), "FL"] },
-      { index: 3, data: ["Houston, James", "Test Corp", renderCities("Dallas", { rowIndex: 3 }), "TX"] },
+      {index: 0, data: {name:"James, Joe", company: "Test Corp", city: renderCities("Yonkers", {rowIndex: 0}),state: "NY"}},
+      {index: 1, data: {name:"Walsh, John", company:"Test Corp", city:renderCities("Hartford", {rowIndex: 1}), state:"CT"}},
+      {index: 2, data: {name:"Herm, Bob", company:"Test Corp", city:renderCities("Tampa", {rowIndex: 2}), state:"FL"}},
+      {index: 3, data: {name:"Houston, James", company:"Test Corp", city:renderCities("Dallas", {rowIndex: 3}), state:"TX"}},
     ];
     renderCities = renderCities;
     renderName = renderName;
   });
 
   it("should render a table", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     assert.strictEqual(
-      shallowWrapper
-        .dive()
-        .dive()
-        .name(),
-      "Paper",
+        shallowWrapper
+            .dive()
+            .dive()
+            .name(),
+        "Paper",
     );
   });
 
   it("should correctly build internal columns data structure", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const actualResult = shallowWrapper.dive().state().columns;
 
     const expectedResult = [
       {
         display: "true",
         name: "Name",
+        field: "name",
         sort: true,
         filter: true,
         download: true,
         sortDirection: null,
         customBodyRender: renderName,
       },
-      { display: "true", name: "Company", sort: true, filter: true, download: true, sortDirection: null },
+      {
+        display: "true", 
+        name: "Company", 
+        field:"company",
+        sort: true, 
+        filter: true, 
+        download: true, 
+        sortDirection: null
+      },
       {
         display: "true",
         name: "City",
+        field: "city",
         sort: true,
         filter: true,
         download: true,
         sortDirection: null,
         customBodyRender: renderCities,
       },
-      { display: "true", name: "State", sort: true, filter: true, download: true, sortDirection: null },
+      {
+        display: "true", 
+        field: "state", 
+        name: "State", 
+        sort: true, 
+        filter: true, 
+        download: true, 
+        sortDirection: null
+      },
     ];
 
     assert.deepEqual(actualResult, expectedResult);
   });
 
   it("should correctly build internal table data and displayData structure", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const state = shallowWrapper.dive().state();
     //assert.deepEqual(state.data, data);
     assert.deepEqual(JSON.stringify(state.displayData), displayData);
   });
 
   it("should correctly re-build internal table data and displayData structure with prop change", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     let state = shallowWrapper.dive().state();
 
     assert.deepEqual(JSON.stringify(state.displayData), displayData);
 
     // now use updated props
-    let newData = data.map(item => [...item]);
-    newData[0][0] = "testing";
-    shallowWrapper.setProps({ data: newData });
+
+    let newData = [
+      {name: "Joe James", company: "Test Corp", city: "Yonkers", state: "NY"},
+      {name: "John Walsh", company: "Test Corp", city: "Hartford", state: "CT"},
+      {name: "Bob Herm", company: "Test Corp", city: "Tampa", state: "FL"},
+      {name: "James Houston", company: "Test Corp", city: "Dallas", state: "TX"}
+    ];
+    
+  /*  let newData = data.map(item => new {}(item));*/
+    newData[0].name = "testing";
+    shallowWrapper.setProps({data: newData});
     shallowWrapper.update();
 
     state = shallowWrapper.dive().state();
     const expectedResult = [
-      { index: 0, data: ["testing", "Test Corp", "Yonkers", "NY"] },
-      { index: 1, data: ["John Walsh", "Test Corp", "Hartford", "CT"] },
-      { index: 2, data: ["Bob Herm", "Test Corp", "Tampa", "FL"] },
-      { index: 3, data: ["James Houston", "Test Corp", "Dallas", "TX"] },
+      {index: 0, data: {name:"testing", company: "Test Corp", city:"Yonkers", state: "NY"}},
+      {index: 1, data: {name:"John Walsh", company:"Test Corp", city:"Hartford", state:"CT"}},
+      {index: 2, data: {name:"Bob Herm", company:"Test Corp", city:"Tampa", state:"FL"}},
+      {index: 3, data: {name:"James Houston", company:"Test Corp", city:"Dallas", state:"TX"}},
     ];
 
     assert.deepEqual(state.data, expectedResult);
@@ -134,7 +162,7 @@ describe("<MaterialDatatable />", function() {
 
   it("should not re-build internal table data and displayData structure with no prop change to data or columns", () => {
     const initializeTableSpy = spy(MaterialDatatable.Naked.prototype, "initializeTable");
-    const mountWrapper = mount(shallow(<MaterialDatatable columns={columns} data={data} />).get(0));
+    const mountWrapper = mount(shallow(<MaterialDatatable columns={columns} data={data}/>).get(0));
 
     let state = mountWrapper.state();
     assert.deepEqual(JSON.stringify(state.displayData), displayData);
@@ -150,7 +178,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should correctly build internal filterList structure", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const state = shallowWrapper.dive().state();
     const expectedResult = [[], [], [], []];
 
@@ -158,7 +186,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should correctly build internal unique column data for filterData structure", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const state = shallowWrapper.dive().state();
     const expectedResult = [
       ["Herm, Bob", "Houston, James", "James, Joe", "Walsh, John"],
@@ -176,7 +204,7 @@ describe("<MaterialDatatable />", function() {
       textLabels,
     };
 
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options}/>);
     const state = shallowWrapper.dive().state();
     assert.strictEqual(state.rowsPerPage, 20);
   });
@@ -186,7 +214,7 @@ describe("<MaterialDatatable />", function() {
       rowsPerPageOptions: [5, 10, 15],
     };
 
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options}/>);
     const state = shallowWrapper.dive().state();
     assert.deepEqual(state.rowsPerPageOptions, [5, 10, 15]);
   });
@@ -196,7 +224,7 @@ describe("<MaterialDatatable />", function() {
       pagination: true,
     };
 
-    const mountWrapper = mount(<MaterialDatatable columns={columns} data={data} options={options} />);
+    const mountWrapper = mount(<MaterialDatatable columns={columns} data={data} options={options}/>);
     const actualResult = mountWrapper.find(MaterialDatatablePagination);
     assert.lengthOf(actualResult, 1);
   });
@@ -206,13 +234,13 @@ describe("<MaterialDatatable />", function() {
       pagination: false,
     };
 
-    const mountWrapper = mount(<MaterialDatatable columns={columns} data={data} options={options} />);
+    const mountWrapper = mount(<MaterialDatatable columns={columns} data={data} options={options}/>);
     const actualResult = mountWrapper.find(MaterialDatatablePagination);
     assert.lengthOf(actualResult, 0);
   });
 
   it("should properly set internal filterList when calling filterUpdate method with type=checkbox", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
     instance.filterUpdate(0, "Joe James", "checkbox");
@@ -223,7 +251,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should remove entry from filterList when calling filterUpdate method with type=checkbox and same arguments a second time", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
     instance.filterUpdate(0, "Joe James", "checkbox");
@@ -240,7 +268,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should properly set internal filterList when calling filterUpdate method with type=dropdown", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
     instance.filterUpdate(0, "Joe James", "dropdown");
@@ -253,13 +281,13 @@ describe("<MaterialDatatable />", function() {
   it("should create Chip when filterList is populated", () => {
     const filterList = [["Joe James"], [], [], []];
 
-    const mountWrapper = mount(<MaterialDatatableFilterList filterList={filterList} filterUpdate={() => true} />);
+    const mountWrapper = mount(<MaterialDatatableFilterList filterList={filterList} filterUpdate={() => true}/>);
     const actualResult = mountWrapper.find(Chip);
     assert.strictEqual(actualResult.length, 1);
   });
 
   it("should remove entry from filterList when calling filterUpdate method with type=dropdown and same arguments a second time", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
     instance.filterUpdate(0, "Joe James", "dropdown");
@@ -277,7 +305,7 @@ describe("<MaterialDatatable />", function() {
 
   it("should properly reset internal filterList when calling resetFilters method", () => {
     // set a filter
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
     instance.filterUpdate(0, "Joe James", "checkbox");
@@ -294,7 +322,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should properly set searchText when calling searchTextUpdate method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />);
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>);
     const table = shallowWrapper.dive();
     const instance = table.instance();
 
@@ -303,14 +331,14 @@ describe("<MaterialDatatable />", function() {
     const state = table.state();
 
     const expectedResult = JSON.stringify([
-      { data: ["James, Joe", "Test Corp", renderCities("Yonkers", { rowIndex: 0 }), "NY"], dataIndex: 0 },
+      {data: ["James, Joe", "Test Corp", renderCities("Yonkers", {rowIndex: 0}), "NY"], dataIndex: 0},
     ]);
 
     assert.deepEqual(JSON.stringify(state.displayData), expectedResult);
   });
 
   it("should sort provided column when calling toggleSortColum method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.toggleSortColumn(0);
@@ -318,17 +346,17 @@ describe("<MaterialDatatable />", function() {
     const state = shallowWrapper.state();
 
     const expectedResult = JSON.stringify([
-      { data: ["Herm, Bob", "Test Corp", renderCities("Tampa", { rowIndex: 0 }), "FL"], dataIndex: 2 },
-      { data: ["Houston, James", "Test Corp", renderCities("Dallas", { rowIndex: 1 }), "TX"], dataIndex: 3 },
-      { data: ["James, Joe", "Test Corp", renderCities("Yonkers", { rowIndex: 2 }), "NY"], dataIndex: 0 },
-      { data: ["Walsh, John", "Test Corp", renderCities("Hartford", { rowIndex: 3 }), "CT"], dataIndex: 1 },
+      {data: ["Herm, Bob", "Test Corp", renderCities("Tampa", {rowIndex: 0}), "FL"], dataIndex: 2},
+      {data: ["Houston, James", "Test Corp", renderCities("Dallas", {rowIndex: 1}), "TX"], dataIndex: 3},
+      {data: ["James, Joe", "Test Corp", renderCities("Yonkers", {rowIndex: 2}), "NY"], dataIndex: 0},
+      {data: ["Walsh, John", "Test Corp", renderCities("Hartford", {rowIndex: 3}), "CT"], dataIndex: 1},
     ]);
 
     assert.deepEqual(JSON.stringify(state.displayData), expectedResult);
   });
 
   it("should toggle provided column when calling toggleViewCol method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.toggleViewColumn(0);
@@ -338,6 +366,7 @@ describe("<MaterialDatatable />", function() {
     const expectedResult = [
       {
         name: "Name",
+        field: "name",
         display: "false",
         sort: true,
         filter: true,
@@ -345,9 +374,18 @@ describe("<MaterialDatatable />", function() {
         sortDirection: null,
         customBodyRender: renderName,
       },
-      { name: "Company", display: "true", sort: true, filter: true, download: true, sortDirection: null },
+      {
+        name: "Company",
+        field: "company",
+        display: "true",
+        sort: true,
+        filter: true,
+        download: true,
+        sortDirection: null
+      },
       {
         name: "City",
+        field: "city",
         display: "true",
         sort: true,
         filter: true,
@@ -355,14 +393,22 @@ describe("<MaterialDatatable />", function() {
         sortDirection: null,
         customBodyRender: renderCities,
       },
-      { name: "State", display: "true", sort: true, filter: true, download: true, sortDirection: null },
+      {
+        name: "State",
+        field: "state",
+        display: "true", 
+        sort: true, 
+        filter: true, 
+        download: true, 
+        sortDirection: null
+      },
     ];
 
     assert.deepEqual(state.columns, expectedResult);
   });
 
   it("should get displayable data when calling getDisplayData method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
     const state = shallowWrapper.state();
 
@@ -371,7 +417,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should update rowsPerPage when calling changeRowsPerPage method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.changeRowsPerPage(10);
@@ -382,7 +428,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should update page position when calling changePage method", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.changePage(2);
@@ -393,7 +439,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should update selectedRows when calling selectRowUpdate method with type=head", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.selectRowUpdate("head", 0);
@@ -401,16 +447,16 @@ describe("<MaterialDatatable />", function() {
 
     const state = shallowWrapper.state();
     const expectedResult = [
-      { index: 0, dataIndex: 0 },
-      { index: 1, dataIndex: 1 },
-      { index: 2, dataIndex: 2 },
-      { index: 3, dataIndex: 3 },
+      {index: 0, dataIndex: 0},
+      {index: 1, dataIndex: 1},
+      {index: 2, dataIndex: 2},
+      {index: 3, dataIndex: 3},
     ];
     assert.deepEqual(state.selectedRows.data, expectedResult);
   });
 
   it("should update selectedRows when calling selectRowUpdate method with type=cell", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.selectRowUpdate("cell", 0);
@@ -421,7 +467,7 @@ describe("<MaterialDatatable />", function() {
   });
 
   it("should update value when calling updateValue method in customBodyRender", () => {
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.updateDataCol(0, 2, "Las Vegas");
@@ -431,8 +477,8 @@ describe("<MaterialDatatable />", function() {
     assert.deepEqual(state.data[0].data[2], "Las Vegas");
   });
   it("should call onTableChange when calling selectRowUpdate method with type=head", () => {
-    const options = { selectableRows: true, onTableChange: spy() };
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options} />).dive();
+    const options = {selectableRows: true, onTableChange: spy()};
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.selectRowUpdate("head", 0);
@@ -440,18 +486,18 @@ describe("<MaterialDatatable />", function() {
 
     const state = shallowWrapper.state();
     const expectedResult = [
-      { index: 0, dataIndex: 0 },
-      { index: 1, dataIndex: 1 },
-      { index: 2, dataIndex: 2 },
-      { index: 3, dataIndex: 3 },
+      {index: 0, dataIndex: 0},
+      {index: 1, dataIndex: 1},
+      {index: 2, dataIndex: 2},
+      {index: 3, dataIndex: 3},
     ];
     assert.deepEqual(state.selectedRows.data, expectedResult);
     assert.strictEqual(options.onTableChange.callCount, 1);
   });
   it("should call onTableChange when calling selectRowUpdate method with type=cell", () => {
-    const options = { selectableRows: true, onTableChange: spy() };
+    const options = {selectableRows: true, onTableChange: spy()};
 
-    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options} />).dive();
+    const shallowWrapper = shallow(<MaterialDatatable columns={columns} data={data} options={options}/>).dive();
     const instance = shallowWrapper.instance();
 
     instance.selectRowUpdate("cell", 0);